Only use ruranges

+++ pyranges/core/pyranges_helpers.py
@@ -38,9 +38,9 @@ def factorize(
     by: VALID_BY_TYPES,
 ) -> ndarray:
     if not by:
-        return np.zeros(len(df), dtype=np.int64)
+        return np.zeros(len(df), dtype=np.uint32)
     _by = arg_to_list(by)
-    return df.groupby(_by).ngroup().to_numpy()
+    return df.groupby(_by).ngroup().to_numpy().astype(np.uint32)
 
 
 def factorize_binary(
@@ -49,9 +49,9 @@ def factorize_binary(
     by: VALID_BY_TYPES,
 ) -> tuple[ndarray, ndarray]:
     if not by:
-        return np.zeros(len(df), dtype=np.int64), np.zeros(len(df2), dtype=np.int64)
+        return np.zeros(len(df), dtype=np.uint32), np.zeros(len(df2), dtype=np.uint32)
     _by = arg_to_list(by)
-    factorized = pd.concat([df[_by], df2[_by]], ignore_index=True).groupby(_by).ngroup().to_numpy()
+    factorized = pd.concat([df[_by], df2[_by]], ignore_index=True).groupby(_by).ngroup().astype(np.uint32).to_numpy()
     return factorized[: len(df)], factorized[len(df) :]
